complete project
this project is about redux..
we can share a state between pages,
for example we have a login name and we can show it in different pages,
our tools are:
store.js that includes state names and functionality & useDispatch and useSelectore hooks.
and now the react 20 mini project is over!